% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su_gfr_taglist.R
\name{su_gfr_taglist}
\alias{su_gfr_taglist}
\title{Generate a snippet of HTML code for summary report}
\usage{
su_gfr_taglist(gfr_elems)
}
\arguments{
\item{gfr_elems}{Google Form Summary Elements object}
}
\description{
Generate a snippet of HTML code for summary report
}
\details{
This will return a list of HTML tags for the elements in gfr_elems
}
\seealso{
\code{\link{su_gfr_elems}}
}
