% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su_meme.R
\name{su_meme}
\alias{su_meme}
\title{Insert a meme}
\usage{
su_meme(
  meme = "yoda-sage",
  height = c("small", "medium", "large")[2],
  quote_main = NULL,
  data_uri = FALSE,
  img_path = "./images/",
  copy_to = NULL,
  tbl_width_px = 640,
  tbl_border_css = "none",
  tbl_margin_css = "0 auto",
  font_size_css = "100\%",
  font_color_css = "black",
  img_boxshadow_css = c("none", "0px 0px 5px 1px #555")[1],
  img_borderradius_css = c("0", "5px")[1],
  debug = FALSE
)
}
\arguments{
\item{meme}{The name of a meme (character)}

\item{height}{The image height (keyword or number)}

\item{quote_main}{The main quote (character)}

\item{data_uri}{Return a URI for the image instead of a file name (logical)}

\item{img_path}{The path to the image (ignored if data_uri is TRUE)}

\item{copy_to}{A directory to copy the image to if needed}

\item{tbl_width_px}{Table width in pixels}

\item{tbl_border_css}{CSS for the table border property}

\item{tbl_margin_css}{CSS for the table margin property}

\item{font_size_css}{CSS for the font-size property}

\item{font_color_css}{CSS for the font color property}

\item{img_boxshadow_css}{CSS for the image box-shadow property}

\item{img_borderradius_css}{CSS for the image border-radius property}

\item{debug}{Print debugging messages to the console, Logical}
}
\value{
html object
}
\description{
Insert a meme
}
\details{
This will return an HTML object that can be printed in a RMarkdown.

\code{height} can be a keyword \code{small|medium|large}, which will make the meme image appear 160, 240, or 320 pixels
tall respectively. Or you can pass an integer. The width will be scaled automatically.
}
\examples{
\dontrun{
su_meme("yoda-sage", "medium", "Organize your work in RStudio Projects,<br/>and peace with your files you will have.")
}

}
