% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su_tips.R
\name{su_tips}
\alias{su_tips}
\title{Generate small HTML files}
\usage{
su_tips(
  gsheet_id,
  gsheet_sheet,
  tips_dir,
  url_base,
  templates_dir = system.file("tips_templates", package = "slideutils"),
  bitly = TRUE,
  open_html = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{gsheet_id}{A Google Spreadsheet ID}

\item{gsheet_sheet}{The name of the sheet in the Google Spreadsheet}

\item{tips_dir}{Directory where to save the generated HTML files}

\item{url_base}{The base of the URL}

\item{templates_dir}{Directory where the HTML template is saved}

\item{bitly}{Create a bitly link, Logical}

\item{open_html}{Open the local HTML file}

\item{quiet}{Suppress messages


This will generate small HTML files that you can use as hints, tips, or solutions for R exercises.
In addition to generating the local HTML file, it will also generate the URL, and optionally a bitly URL,
that you can embed in your .Rmd or .R file.

The way this works is you have a spreadsheet with columns that contain the content of the questions. Currently,
only Google Sheets are supported. You can model your spreadsheet after the following:

https://docs.google.com/spreadsheets/d/1iEOGHh4hC4KbiD5pn4kBTBMt3zrJga-rvSilR98ei00/copy

The spreadsheet does not have to be open to anyone with the link. You'll be prompted to for your credentials using
the googlesheets4 package.

This script goes through the sheet and
   i) renders (creates HTML file) for all rows where 'needs_rendering' is blank or TRUE
   ii) msg_html will be injected into a <div>
  iii) updates the Google Sheet
  needs_rendering - set to FALSE
  url - populated
  bitly - populated}
}
\value{
A vector of HTML files rendered
}
\description{
Generate small HTML files based on content from a spreadsheet
}
\examples{
\dontrun{

}

}
